+++ cmd/lambda/update_principal_policy/main.go
@@ -61,6 +61,7 @@ func handler(ctx context.Context, snsEvent events.SNSEvent) error {
 			PrincipalRoleName:    common.RequireEnv("PRINCIPAL_ROLE_NAME"),
 			PrincipalPolicyName:  common.RequireEnv("PRINCIPAL_POLICY_NAME"),
 			PrincipalIAMDenyTags: strings.Split(common.RequireEnv("PRINCIPAL_IAM_DENY_TAGS"), ","),
+			Regions:              strings.Split(common.RequireEnv("ALLOWED_REGIONS"), ","),
 			PolicyBucket:         common.RequireEnv("ARTIFACTS_BUCKET"),
 			PolicyBucketKey:      common.RequireEnv("PRINCIPAL_POLICY_S3_KEY"),
 		})
@@ -78,6 +79,7 @@ type processRecordInput struct {
 	PrincipalRoleName    string
 	PrincipalPolicyName  string
 	PrincipalIAMDenyTags []string
+	Regions              []string
 	PolicyBucket         string
 	PolicyBucketKey      string
 }