+++ modules/variables.tf
@@ -17,11 +17,6 @@ variable "namespace" {
   description = "The namespace for this Terraform run"
 }
 
-variable "organization_id" {
-  description = "The AWS Orgnanization ID the AWS Account is under."
-  default     = "STUB"
-}
-
 variable "reset_nuke_template_bucket" {
   description = "S3 bucket name containing the nuke configuration template. Use this to override the default nuke configuration."
   default     = "STUB"
@@ -54,7 +49,7 @@ variable "reset_image_pull_creds" {
 
 variable "reset_nuke_toggle" {
   description = "Indicator to set Nuke to not delete any resources. Use 'false' to indicate a Dry Run. NOTE: Cannot change Account status with this toggled off."
-  default     = "false"
+  default     = "true"
 }
 
 variable "populate_reset_queue_schedule_expression" {
@@ -78,7 +73,8 @@ variable "check_budget_enabled" {
   description = "If false, budgets will not be checked"
 }
 variable "budget_notification_from_email" {
-  type = string
+  type    = string
+  default = "notifications@example.com"
 }
 
 variable "budget_notification_bcc_emails" {
@@ -198,4 +194,29 @@ variable "principal_budget_period" {
   type        = string
   description = "Principal budget period must be WEEKLY or MONTHLY"
   default     = "WEEKLY"
+}
+
+variable "allowed_regions" {
+  type = list(string)
+  default = [
+    "us-east-2",
+    "us-east-1",
+    "us-west-1",
+    "us-west-2",
+    "ap-east-1",
+    "ap-south-1",
+    "ap-northeast-3",
+    "ap-northeast-2",
+    "ap-southeast-1",
+    "ap-southeast-2",
+    "ap-northeast-1",
+    "ca-central-1",
+    "eu-central-1",
+    "eu-west-1",
+    "eu-west-2",
+    "eu-west-3",
+    "eu-north-1",
+    "me-south-1",
+    "sa-east-1"
+  ]
 }