+++ pkg/usage/usage.go
@@ -23,7 +23,9 @@ type DB struct {
 	// DynamoDB Client
 	Client *dynamodb.DynamoDB
 	// Name of the Usage table
-	UsageTableName string
+	UsageTableName   string
+	PartitionKeyName string
+	SortKeyName      string
 	// Use Consistend Reads when scanning or querying.  When possbile.
 	ConsistendRead bool
 }
@@ -44,6 +46,7 @@ type Usage struct {
 type Service interface {
 	PutUsage(input Usage) error
 	GetUsageByDateRange(startDate time.Time, endDate time.Time) ([]*Usage, error)
+	GetUsageByPrincipal(startDate time.Time, principalID string) ([]*Usage, error)
 }
 
 // PutUsage adds an item to Usage DB
@@ -126,13 +129,64 @@ func (db *DB) GetUsageByDateRange(startDate time.Time, endDate time.Time) ([]*Us
 	return usageRecords, nil
 }
 
+// GetUsageByPrincipal returns usage amount for all leases for input Principal
+// startDate is epoch Unix date
+func (db *DB) GetUsageByPrincipal(startDate time.Time, principalID string) ([]*Usage, error) {
+
+	output := make([]*Usage, 0)
+
+	// Convert startDate to the start time of that day
+	usageStartDate := time.Date(startDate.Year(), startDate.Month(), startDate.Day(), 0, 0, 0, 0, time.UTC)
+	currentDate := time.Now()
+	usageEndDate := time.Date(currentDate.Year(), currentDate.Month(), currentDate.Day(), 23, 59, 59, 0, time.UTC)
+
+	if usageEndDate.Sub(usageStartDate) < 0 {
+		errorMessage := fmt.Sprintf("UsageStartDate \"%d\" should be a past date or today's date \"%d\".", usageStartDate.Unix(), usageEndDate.Unix())
+		log.Print(errorMessage)
+		return nil, nil
+	}
+
+	for {
+
+		var resp, err = db.Client.GetItem(getInputForGetUsageByPrincipalID(db, usageStartDate, principalID, db.ConsistendRead))
+		if err != nil {
+			errorMessage := fmt.Sprintf("Failed to query usage record for start date \"%s\": %s.", startDate, err)
+			log.Print(errorMessage)
+			return nil, err
+		}
+
+		item := Usage{}
+
+		err = dynamodbattribute.UnmarshalMap(resp.Item, &item)
+		if err != nil {
+			errorMessage := fmt.Sprintf("Failed to unmarshal Record, %v", err)
+			log.Print(errorMessage)
+			return nil, err
+		}
+
+		output = append(output, &item)
+
+		// increment startdate by a day
+		usageStartDate = usageStartDate.AddDate(0, 0, 1)
+
+		// continue to get usage till usageEndDate
+		if usageEndDate.Sub(usageStartDate) < 0 {
+			break
+		}
+	}
+
+	return output, nil
+}
+
 // New creates a new usage DB Service struct,
 // with all the necessary fields configured.
-func New(client *dynamodb.DynamoDB, usageTableName string) *DB {
+func New(client *dynamodb.DynamoDB, usageTableName string, partitionKeyName string, sortKeyName string) *DB {
 	return &DB{
-		Client:         client,
-		UsageTableName: usageTableName,
-		ConsistendRead: false,
+		Client:           client,
+		UsageTableName:   usageTableName,
+		PartitionKeyName: partitionKeyName,
+		SortKeyName:      sortKeyName,
+		ConsistendRead:   false,
 	}
 }
 
@@ -154,6 +208,8 @@ func NewFromEnv() (*DB, error) {
 			aws.NewConfig().WithRegion(common.RequireEnv("AWS_CURRENT_REGION")),
 		),
 		common.RequireEnv("USAGE_CACHE_DB"),
+		"StartDate",
+		"PrincipalId",
 	), nil
 }
 
@@ -187,4 +243,21 @@ func getQueryInput(tableName string, startDate time.Time, startKey map[string]*d
 		},
 		ConsistentRead: aws.Bool(consistentRead),
 	}
+}
+
+// getInputForGetUsageByPrincipalID returns a GetItemInput for given inputs
+func getInputForGetUsageByPrincipalID(d *DB, startDate time.Time, principalID string, consistentRead bool) *dynamodb.GetItemInput {
+	getItemInput := dynamodb.GetItemInput{
+		TableName: aws.String(d.UsageTableName),
+		Key: map[string]*dynamodb.AttributeValue{
+			d.PartitionKeyName: {
+				N: aws.String(strconv.FormatInt(startDate.Unix(), 10)),
+			},
+			d.SortKeyName: {
+				S: aws.String(principalID),
+			},
+		},
+		ConsistentRead: aws.Bool(consistentRead),
+	}
+	return &getItemInput
 }